+++ src/renderer/app.module.ts
@@ -5,7 +5,6 @@ import { HttpModule } from '@angular/http';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { DatePipe, APP_BASE_HREF } from '@angular/common';
 import { ColorPickerModule } from 'ngx-color-picker';
-import { DynamicHTMLModule } from 'ng-dynamic';
 
 import * as Components from './components';
 import * as SvgComponents from './svg-components';
@@ -15,30 +14,13 @@ import * as Pipes from './pipes';
 import * as Guards from './guards';
 import { AppRoutes } from './app.routing';
 
-function ngObjectsToArray(importObject: any, selector: boolean = false) {
-    if (selector === true) {
-        let objectArray: { component: any, selector: string }[] = [];
-        for (let attribute in importObject) {
-            if (typeof importObject[attribute] === 'function') {
-                let metadata = Reflect.getMetadata('annotations', importObject[attribute]);
-                for (let i = 0; i < metadata.length; i++) {
-                    if (metadata[i].selector) {
-                        objectArray.push({ component: importObject[attribute], selector: metadata[i].selector });
-                        break;
-                    }
-                }
-            }
-        }
-        return objectArray;
+function ngObjectsToArray(importObject: any) {
+    let objectArray: any[] = [];
+    for (let attribute in importObject) {
+        if (typeof importObject[attribute] === 'function')
+            objectArray.push(importObject[attribute]);
     }
-    else {
-        let objectArray: any[] = [];
-        for (let attribute in importObject) {
-            if (typeof importObject[attribute] === 'function')
-                objectArray.push(importObject[attribute]);
-        }
-        return objectArray;
-    }
+    return objectArray;
 }
 
 @NgModule({
@@ -49,12 +31,7 @@ function ngObjectsToArray(importObject: any, selector: boolean = false) {
         AppRoutes,
         FormsModule,
         ReactiveFormsModule,
-        ColorPickerModule,
-        DynamicHTMLModule.forRoot({
-            components: [].concat(
-                ngObjectsToArray(SvgComponents, true)
-            )
-        })
+        ColorPickerModule
     ],
     declarations: [].concat(
         ngObjectsToArray(Components),